PowerSimulationsDynamicsSurrogates

Documentation for PowerSimulationsDynamicsSurrogates.

PowerSimulationsDynamicsSurrogates.SteadyStateNODE

InfrastructureSystems.get_internal

InfrastructureSystems.get_name

PowerSimulationsDynamicsSurrogates.create_subsystem_from_buses

PowerSimulationsDynamicsSurrogates.generate_train_data

PowerSimulationsDynamicsSurrogates.get_exogenous_bias

PowerSimulationsDynamicsSurrogates.get_exogenous_scale

PowerSimulationsDynamicsSurrogates.get_initializer_parameters

PowerSimulationsDynamicsSurrogates.get_initializer_structure

PowerSimulationsDynamicsSurrogates.get_node_parameters

PowerSimulationsDynamicsSurrogates.get_node_structure

PowerSimulationsDynamicsSurrogates.get_observer_parameters

PowerSimulationsDynamicsSurrogates.get_observer_structure

PowerSimulationsDynamicsSurrogates.get_x_bias

PowerSimulationsDynamicsSurrogates.get_x_scale

PowerSystems.get_base_power

PowerSystems.get_ext

PowerSystems.get_n_states

PowerSystems.get_states

PowerSystems.set_base_power!

PowerSystems.set_ext!

PowerSimulationsDynamicsSurrogates.SteadyStateNODE — Type

mutable struct SteadyStateNODE <: DynamicInjection
    name::String
    initializer_structure::Vector{Tuple{Int64, Int64, Bool, String}}
    initializer_parameters::Vector{Float64}
    node_structure_exogenous::Vector{Tuple{Int64, Int64, Bool, String}}
    node_structure_states::Vector{Tuple{Int64, Int64, Bool, String}}
    node_structure_common::Vector{Tuple{Int64, Int64, Bool, String}}
    node_parameters::Vector{Float64}
    observer_structure::Vector{Tuple{Int64, Int64, Bool, String}}
    observer_parameters::Vector{Float64}
    x_scale::Vector{Float64}
    x_bias::Vector{Float64}
    exogenous_scale::Vector{Float64}
    exogenous_bias::Vector{Float64}
    base_power::Float64
    states::Vector{Symbol}
    n_states::Int
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

Experimental surrogate

Arguments

name::String

initializer_structure::Vector{Tuple{Int64, Int64, Bool, String}}: layers of the initializer

initializer_parameters::Vector{Float64}: parameters of the initializer

node_structure_exogenous::Vector{Tuple{Int64, Int64, Bool, String}}: layers of the initializer

node_structure_states::Vector{Tuple{Int64, Int64, Bool, String}}: layers of the initializer

node_structure_common::Vector{Tuple{Int64, Int64, Bool, String}}: layers of the initializer

node_parameters::Vector{Float64}: parameters of the node

observer_structure::Vector{Tuple{Int64, Int64, Bool, String}}: layers of the initializer

observer_parameters::Vector{Float64}: parameters of the node

x_scale::Vector{Float64}: scale for x input

x_bias::Vector{Float64}: bias for x input

exogenous_scale::Vector{Float64}: scale for exogenous input

exogenous_bias::Vector{Float64}: bias for exogenous input

base_power::Float64: Base power

states::Vector{Symbol}: The states of GenericDER depend on the Flags

n_states::Int: The states of GenericDER depend on the Flags

ext::Dict{String, Any}

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_internal — Method

Get SteadyStateNODE internal.

InfrastructureSystems.get_name — Method

Get SteadyStateNODE name.

PowerSimulationsDynamicsSurrogates.create_subsystem_from_buses — Method

createsubsystemfrombuses(sysfull::PSY.System, subsystembusnumbers::Vector{Int64}) -Takes a deepcopy of PSY system and removes all components that are not attached to the buses included in subsystem_bus_numbers. -Leaves all branches that connect two buses within the surrogate AND branches that connect to the rest of the system ("connecting buses") -Adds source components at connecting buses –- these are used to perturb the surrogate when generating training data.

PowerSimulationsDynamicsSurrogates.generate_train_data — Method

function generate_train_data(
    sys_train::PSY.System,
    perturbations::Vector{SurrogatePerturbation},
    operating_points::Vector{SurrogateOperatingPoint},
    data_name::String,   
    data_collection_params::GenerateDataParams,
)

Options for data_name: "SteadyStateNODEData"

PowerSimulationsDynamicsSurrogates.get_exogenous_bias — Method

Get SteadyStateNODE exogenous_bias.

PowerSimulationsDynamicsSurrogates.get_exogenous_scale — Method

Get SteadyStateNODE exogenous_scale.

PowerSimulationsDynamicsSurrogates.get_initializer_parameters — Method

Get SteadyStateNODE initializer_parameters.

PowerSimulationsDynamicsSurrogates.get_initializer_structure — Method

Get SteadyStateNODE initializer_structure.

PowerSimulationsDynamicsSurrogates.get_node_parameters — Method

Get SteadyStateNODE node_structure_states.

PowerSimulationsDynamicsSurrogates.get_node_structure — Method

Get SteadyStateNODE node_structure_exogenous.

PowerSimulationsDynamicsSurrogates.get_observer_parameters — Method

Get SteadyStateNODE observer_parameters.

PowerSimulationsDynamicsSurrogates.get_observer_structure — Method

Get SteadyStateNODE observer_structure.

PowerSimulationsDynamicsSurrogates.get_x_bias — Method

Get SteadyStateNODE x_bias.

PowerSimulationsDynamicsSurrogates.get_x_scale — Method

Get SteadyStateNODE x_scale.

PowerSystems.get_base_power — Method

Get SteadyStateNODE base_power.

PowerSystems.get_ext — Method

Get SteadyStateNODE ext.

PowerSystems.get_n_states — Method

Get SteadyStateNODE n_states.

PowerSystems.get_states — Method

Get SteadyStateNODE states.

PowerSystems.set_base_power! — Method

Set SteadyStateNODE base_power.

PowerSystems.set_ext! — Method

Set SteadyStateNODE ext.